mmc: generate the interrupt unexpected
When host generated the CMD conflicts interrupts, driver needs a small
delay before clearing it otherwise such interrupts will be generated
again after clearing it. This behavior is not described in SDHC spec,
and suspect it is a silicon issue.

Before confirm whether this is a silicon issue, have a small delay is
still an avaliable software workaround

Fixed conflict porting from 3.10 to 4.x

Signed-off-by: Ning Feiyi <feiyix.ning@intel.com>
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Conflicts:
	drivers/mmc/host/sdhci.c

Change-Id: I09cb4faaf712efbc1ce1982e3ef08b0a90ebce10