[REVERTME] sdhci-pci: power up Merrifield SDHCI hosts
After enabling D0ix on sdio0/1/emmc0 and rebooting the platform

The issue is, emmc0 is in d0i3 during reboot and there is a access
without bringing the device to d0i0.

Fixed this by implementing shutdown handler and disabling runtime_pm
during shutdown, this is similar to other sdhci driver implementations.

However this solution is a workaround since during shutdown emmc driver
is not expecting access, and since there is access need to find who
is accessing and fix that issue.

Till that time we have this solution to unblock S0ix enabling.

This patch also includes:
 bd46f3f8 [REVERTME] sdhci-pci: implement shutdown callback

Fixed conflict porting 3.10 to 4.x; aswell as add the .shutdown
callback for struct pci driver that is not present in upstream
(might need to be removed as this is a revertme patch, but no
other patch is in the current 3.10 tree that does the revert)

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Conflicts:
	drivers/mmc/host/sdhci-pci.c

Change-Id: I366466c73d16d592513f91af9d7b27c70ea8c4db