platform_wifi: Add support for Fast IRQ
In case where the Fast IRQ mode is enabled in the SFI table, we make use of it
and change the default IRQ flags associated with the OOB interrupt. This is
because we are using level interrupts when in fast IRQ mode.

Fixed conflicts from 3.10 port to 4.x upstream (quirk variance and no using iwlwifi)
Conflicts:
	arch/x86/platform/intel-mid/board.c
	arch/x86/platform/intel-mid/device_libs/platform_wifi.c

Change-Id: I423a95abd971ee2e53fde1ac63e1983e8c40b14c