PTI driver: add support of Merrifield platform
This patch corresponds to the porting of a patch that had not made
its way to kernel 3.10. This change consists in supporting Intel
Merrifield platform. A generic mechanism to support PTI device
options is introduced.

Initial change had already been ported to kernel 3.4 with
84a017e: PTI: Ported PTI patches from JB main tree (http://android.intel.com:8080/74298)

Fixed conflict porting 3.10 to 4.x

Signed-off-by: Yann Fouassier <yann.fouassier@intel.com>

Conflicts:
	drivers/misc/pti.c

Change-Id: I9899cb828004e5abdabeab194ea1705f6cac3b96