usb/dwc3-device: transfer 0-length packet to satisfy host request
The current driver may have some issue if the control transfer
length of the data stage is multiple of max packet size.

This patch provides a fix by issuing another Start Transfer in
case that host is trying to move an extra 0-length packet.

Fixed conflict porting 3.10 to 4.x

Signed-off-by: Jiebing Li <jiebing.li@intel.com>
Signed-off-by: Wang, Yu <yu.y.wang@intel.com>

Conflicts:
	drivers/usb/dwc3/ep0.c

Change-Id: I5f00b21ceb0a16b6550584ac40e915bb82d82160