Check for NC and SC device status before entering S3
During S3 flow there could be a possibility that S0ix aborts happens.
As seen in BZ 114607. But today due to the lack of s0ix abort feedback
mechanism the platform remains in kind of unknown state until a power
button or USB is connected. This gap is addressed in BZ 12353.

This patch adds a check in PMU driver to check NC and SC device status
before requesting C7 (basically C6 mwait + s0ix hint). S3 flow is aborted
in case the check fails and retries; rather leaving platform in unknown
state.

Signed-off-by: Ananth Krishna <ananth.krishna.r@intel.com>
[Fix some trivial conflicts]
Signed-off-by: Srinidhi Kasagar <srinidhi.kasagar@intel.com>