EDISON: Remove SPID
The SPID mechanism is not needed for EDISON.
Actually, better not to have it at all to simplify the code
and avoid the need of provisionning SPID on manufacturing line.

Remove the SPID implementation and all references to SPID macros in the code.

Fixed conflict porting 3.10 to 4.x (commit will be dropped for Edison along with previous commits that added the SPID handling code in the first place).

Conflicts:
	arch/x86/platform/intel-mid/device_libs/platform_hsu.c
	arch/x86/platform/intel-mid/device_libs/platform_mid_pwm.c
	arch/x86/platform/intel-mid/device_libs/platform_mrfl_thermal.c
	arch/x86/platform/intel-mid/device_libs/platform_msic_thermal.c
	arch/x86/platform/intel-mid/device_libs/platform_soc_thermal.c
	arch/x86/platform/intel-mid/intel_mid_sfi.c
	arch/x86/platform/intel-mid/spid.c

Change-Id: I2febcd01cbf0570511cbc395d86e8dcc62d3426f