Add PMIC and DESIGNWARE i2c drivers
We import the i2c designware code from external_drivers directory.
We also added the i2c-pmic which is mandatory to probe the bq24261
driver.

Reworked/Fixed conflicts for 3.10 to 4.x porting,
this still needs a lot of work before it could ever go upstream
for merrifield like baytrail, etc... that is already upstream.

Signed-off-by: Loïc Akue <loicx.akue@intel.com>

Conflicts:
	drivers/i2c/busses/i2c-designware-core.c
	drivers/i2c/busses/i2c-designware-core.h
	drivers/i2c/busses/i2c-designware-pcidrv.c
	drivers/i2c/busses/i2c-designware-platdrv.c

Change-Id: I768442be4583b8e92d0cbf9a92821b03d16ccf94