spi: do not toggle Chip Select while transferring a message
In polling mode, each word are sent after verifying the output FIFO is empty.
This behavior makes the CS line automatically de-asserted between each word by
the SPI Core. That is not SPI compliant.

This patch makes the CS handled as a gpio and not automatically by the SPI
Core.

Change-Id: I693eca4adc974fbdb2e7879d1075212b8146ba93
Signed-off-by: Simon Desfarges <simonx.desfarges@intel.com>