spi: allow runtime suspend after 25ms delay
When the spi driver is sleeping (D0i3), the framework takes 10-20ms to wake it
up at each new message (ioctl). This behavior prevents the user issuing ioctls
on spi device at a rate higher than 50Hz.

This patch adds a delay between the end of the transfer and the runtime
suspend, so the device does not suspend after each transfer.

The ioctl issuing rate raises up to 2KHz.

Signed-off-by: Simon Desfarges <simonx.desfarges@intel.com>