audio: add a dummy machine driver
Even if there is no audio codec plugged to the Edison board,
we want to be able to control the Audio DSP to allow the use
of the SSP1 and SSP2 ports.

This patch aims to create a dummy machine driver using a dummy codec
to allow a user to access the Audio DSP alsa controls.

Fixed conflict porting 3.10 to 4.x

Signed-off-by: Michael Soares <michaelx.soares@intel.com>

Conflicts:
	sound/soc/intel/board/Makefile

Change-Id: Ibc32e56c23d69d95b07ec325656952d70ed81783