MMC hang during reboot from dequeue of mmc device
This commit resolves the issue where the Intel Edison
will hang if rebooted more than once (repeated reboots).
It was found that the mmc device was being registerd
twice with 2 shutdown functions. When a device
reboot/shutdown operation was called the same mmc device
was trying to be cancelled from a workqueue. This
situation caused a race condition on the second operation
since the cancelled work will wait for completion.

Also added in code needed for proper resume setting for MMC
power state.

Fixed conflict porting 3.10 to 4.x

Also, might need to add this code change later when bcmdhd
4.x version of Broadcom is added:

--- a/drivers/staging/edison-bcm43340/bcmsdh_sdmmc_linux.c
+++ b/drivers/staging/edison-bcm43340/bcmsdh_sdmmc_linux.c
@@ -362,6 +362,8 @@ static int bcmsdh_sdmmc_resume(struct device *pdev)
 #if defined(OOB_INTR_ONLY)
        bcmsdh_resume(sdioh->bcmsdh);
 #endif
+       if (func->card && func->card->host)
+               func->card->host->pm_flags &= ~MMC_PM_KEEP_POWER;

        smp_mb();

Change-Id: I1400df84e178d2b582725db2129ea7ee7c28a4da
Signed-off-by: Brian Wood <brian.j.wood@intel.com>

Conflicts:
	drivers/mmc/card/block.c
	drivers/mmc/core/mmc.c
	drivers/staging/edison-bcm43340/bcmsdh_sdmmc_linux.c