usb: gadget: configfs: Let composite_setup handle setup requests
android_setup() iterates over the list of available function drivers
attempting to find one that will handle the setup request.  This has
several issues: 1) it's the list of available functions, not the ones
actually used in the current configuration, 2) the corresponding
function structures may not have been setup yet, or may have been freed
already, and 3) access to the list must be protected by the gi->lock
mutex.  Since composite_setup() already does the right thing here by
calling the setup() callback for functions that are actually part of the
configuration (with the proper locking), kill off the loop over the
available_func list.

Change-Id: I69dc69da173a652b3702ef738da65f5684310a44
Fixes: bb2a9e71710e ("usb: gadget: Add Uevent to notify userspace")
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/309900
Reviewed-by: Badhri Jagan Sridharan <badhri@google.com>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Signed-off-by: Mian Yousaf Kaukab <yousaf.kaukab@intel.com>