usb: gadget: Fix handling of unrecognized OS descriptor requests
If composite_setup() receives a control request for an OS Feature
Descriptor that is not Extended Compatibility ID or Extended Properties,
it falls through to enqueueing a response of length 'value' bytes which,
because the request is not recognized, is -EOPNOTSUPP.  Fix it so that
we instead bail out of composite_setup() without enqueueing a response.
The UDC driver is then responsible for halting EP0.

Change-Id: I8b94bc59b66cd7258e502355737a409f5029a95b
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/313224
Reviewed-by: Benson Leung <bleung@chromium.org>
Signed-off-by: Mian Yousaf Kaukab <yousaf.kaukab@intel.com>