gpio-intel-mid: Program internal pullup/pulldown functionality
Commit based on GPIO output pins value to resolve contention
and thereby save power

The GPIO pins that are configured and used as output pins, should have
an internal pullup/pulldown based on the logic that particular output
pin is driving at that moment.
Say if a GPIO output pin is driving logic low, and if there is a pullup,
then there will be contention - wherein there will be a clear path of
current to leak. And if neither pullup nor pulldown is present/configured,
there will still be leakage across SoC. Hence program the pullup or pulldown
dynamically based on the value that output pin is driving.
Also I2C pins are untouched with this patch.

Change-Id: I2d060b6a26d5985ef7b70323a3c2b71711b4b121
Signed-off-by: Brian Wood <brian.j.wood@intel.com>