sdhci-pci-core: power up Merrifield SDHCI hosts
After enabling D0ix on sdio0/1/emmc0 and rebooting the platform

The issue is, emmc0 is in d0i3 during reboot and there is a access
without bringing the device to d0i0.

Fixed this by implementing shutdown handler and disabling runtime_pm
during shutdown, this is similar to other sdhci driver implementations.

Fixed conflict porting 3.10 to 4.x as well as add the .shutdown
callback for struct pci driver.

Change-Id: If9c275f8c747dc56d813668de53658c28b49775e
Signed-off-by: Brian Wood <brian.j.wood@intel.com>