MMC hang during reboot from dequeue of mmc device
This commit resolves the issue where the Intel Edison
will hang if rebooted more than once (repeated reboots).
It was found that the mmc device was being registerd
twice with 2 shutdown functions. When a device
reboot/shutdown operation was called the same mmc device
was trying to be cancelled from a workqueue. This
situation caused a race condition on the second operation
since the cancelled work will wait for completion.

Code was added in for proper resume setting for MMC
power state.

Also, might need to add this code change later when bcmdhd
4.x version of Broadcom wifi driver support is added
to this file:
	drivers/staging/edison-bcm43340/bcmsdh_sdmmc_linux.c
at function:
	 static int bcmsdh_sdmmc_resume(struct device *pdev)

 #if defined(OOB_INTR_ONLY)
        bcmsdh_resume(sdioh->bcmsdh);
 #endif
+       if (func->card && func->card->host)
+               func->card->host->pm_flags &= ~MMC_PM_KEEP_POWER;

        smp_mb();

Change-Id: I5dcaadc754bb6a5926689c0917c87dce5740e152
Signed-off-by: Brian Wood <brian.j.wood@intel.com>