wifi: broadcom: Fix wifi wowlan 80211 kernel crash
In the 4.x+ kernel there is an issue with the Broadcom wifi driver
configuration of the wowlan_config struct and its use by the nl80211*
functions where a kernel crash can result. This commit resolves the
issue by letting the nl80211 driver setup the defaults so it can do the
alloc/free operations.

Change-Id: I7a040f2ae3eeb0a6e6aedaa97060281ef39ab1ae
Signed-off-by: Brian Wood <brian.j.wood@intel.com>